backend/requirements.txt
crawl4ai>=0.1.0
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0 